stanovanja/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.9.3
black==24.8.0
boto3==1.35.35
botocore==1.35.35
certifi==2024.8.30
chardet==4.0.0
Django==3.2.25
django-filter==2.4.0
django-modelcluster==5.3
django-storages==1.14.4
django-taggit==1.5.1
django-treebeard==4.7.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==5.13.2
jmespath==1.0.1
l18n==2021.3
openpyxl==3.1.5
Pillow==8.4.0
psycopg2==2.8.6
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.25.1
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.6.1
telepath==0.3.1
typing-extensions==4.12.2
urllib3==1.26.20
wagtail==2.14.2
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
